Create tests for projectile and damage
Created tests to ensure after every skill attack, damage is done to the enemy. And when the enemy attacks that damage is done to the player.

Also wrote tests for projectiles testing that it's produced with every player skill call moving towards the enemy and it is destroyed when the skill hits the enemy.